Refactor MentionInput
