--- demo/components.css
@@ -1,15 +1,0 @@
-.ProseMirror-selectednode > div {
-  outline: 2px solid #8cf;
-}
-
-.chip-container {
-  display: inline-block;
-  padding: 4px 2px 4px 0;
-}
-
-.ProseMirror {
-  p {
-    align-items: center;
-    flex-wrap: wrap;
-  }
-}
